cnbn/requirements.txt
scikit-learn==1.3.2
nltk==3.8.1
pandas==2.0.0
guessLang==2.2.1
PyGithub